refactor: broke down encoding/decoding logic into separate utility files
